Change maxCycle argument in Agent.learn method; int -> long

+++ src/jp/ac/nara_k/info/tetris_4w_ren/agent/Agent.java
@@ -37,8 +37,8 @@ public abstract class Agent {
         System.out.println(environment);
     }
 
-    public void learn(int maxCycle) {
-        for (int cycle = 1; cycle<= maxCycle; cycle++) {
+    public void learn(long maxCycle) {
+        for (long cycle = 1; cycle<= maxCycle; cycle++) {
             doCycle();
             initEnvironment();
         }
